changed setup to export the images with the right image description for srcset

+++ app/Livewire/LandingPage.php
@@ -6,7 +6,6 @@ use Livewire\Component;
 use Intervention\Image\ImageManager;
 use Intervention\Image\Drivers\Gd\Driver;
 use Livewire\WithFileUploads;
-use Intervention\Image\Encoders\JpegEncoder;
 
 class LandingPage extends Component
 {
@@ -18,6 +17,7 @@ class LandingPage extends Component
     public $availableWidths = [];
     public $download = false;
     public $nom = '';
+    public $extension = '';
 
     protected $rules = [
         'photo' => 'required|image|max:20400',
@@ -30,87 +30,90 @@ class LandingPage extends Component
 
         if ($this->imageInfo) {
             $imageWidth = $this->imageInfo[0];
-            $this->availableWidths = array_filter($this->widths, function($width) use ($imageWidth) {
+            $this->widths = array_filter($this->widths, function($width) use ($imageWidth) {
                 return $width <= $imageWidth;   });
         }
     }
 
     public function scaleImage(){
         $this->validate($this->rules);
-        $this->nom =  $this->photo->getClientOriginalName();
-        $extension = $this->photo->getClientOriginalExtension();
-
-        $this->photo->storeAs('photos', $this->nom, 'public'); 
-        
-        if($extension == 'png'){
-            //Convert to .jpeg, store image and destroy gd file to free memory
-            $image_gd = @imagecreatefrompng(storage_path('app/public/photos/'.$this->nom));
-            imagejpeg($image_gd, storage_path('app/public/photos/'.$this->nom ));
+        $this->nom = pathinfo($this->photo->getClientOriginalName(), PATHINFO_FILENAME);
+        $this->extension = $this->photo->getClientOriginalExtension();
+    
+        $this->photo->storeAs('photos', $this->nom . '.' . $this->extension, 'public'); 
+    
+        if ($this->extension == 'png') {
+            // Convert to .jpeg, store image, and destroy gd file to free memory
+            $image_gd = @imagecreatefrompng(storage_path('app/public/photos/' . $this->nom . '.' . $this->extension));
+            imagejpeg($image_gd, storage_path('app/public/photos/' . $this->nom . '.jpg'));
             imagedestroy($image_gd);
+            $this->extension = 'jpg';
         }
-
-        try{
+    
+        try {
             $manager = new ImageManager(new Driver());
-            $image = $manager->read(storage_path('app/public/photos/'.$this->nom ));
+            $image = $manager->read(storage_path('app/public/photos/' . $this->nom . '.' . $this->extension));
             
-            if($this->availableWidths){
-                foreach( $this->availableWidths as $width){
+            if ($this->availableWidths) {
+                foreach ($this->availableWidths as $width) {
                     $resizedImage = clone $image;
                     $resizedImage->scale(width: $width);
-                    $resizedImage->save(storage_path('app/public/photos_edited/'.$width.'_'.$this->nom));
-                }}
-
+                    $resizedImageName = $this->nom . '-' . $width . 'w.' . $this->extension;
+                    $resizedImage->save(storage_path('app/public/photos_edited/' . $resizedImageName));
+                }
+            }
+    
             $this->statut = 'Toutes les images ont été redimensionnées avec succès !';
             $this->download = true;
-
-        }catch(\Exception $e){
-            dump( $e->getMessage());
+    
+        } catch (\Exception $e) {
+            dump($e->getMessage());
         }
     }
+    
 
     public function downloadImages(){
-
-        if($this->availableWidths){
+        if ($this->availableWidths) {
             if (sizeof($this->availableWidths) > 1) {
                 $zip = new \ZipArchive();
-                $zipFileName = 'images_' . time() . '.zip';
+                $zipFileName = 'ImageSizeify-'.$this->nom.'-'.time().'.zip';
                 $zipFilePath = storage_path('app/public/photos_edited/' . $zipFileName);
-            
+    
                 if ($zip->open($zipFilePath, \ZipArchive::CREATE | \ZipArchive::OVERWRITE) === TRUE) {
                     foreach ($this->availableWidths as $width) {
-                        $filePath = storage_path('app/public/photos_edited/' . $width . '_' . $this->nom);
+                        $resizedImageName = $this->nom . '-' . $width . 'w.' .$this->extension;
+                        $filePath = storage_path('app/public/photos_edited/' .$resizedImageName);
                         if (file_exists($filePath)) {
                             $zip->addFile($filePath, basename($filePath));
                         }
                     }
                     $zip->close();
-                } else {
-                    return response()->json(['error' => 'Erreur de création de fichier zip'], 500);
-                }
-
-                //Supprimer l'image original
-                $originalImg = storage_path('app/public/photos/'.$this->nom);
+                } else {    return response()->json(['error' => 'Erreur de création de fichier zip'], 500);}
+    
+                // Delete the original image
+                $originalImg = storage_path('app/public/photos/'.$this->nom.'.'.$this->extension);
                 if (file_exists($originalImg)) {
-                    unlink($originalImg);
-                }
-
-                //Supprimer les images après création du fichier zip
+                    unlink($originalImg);   }
+    
+                // Delete the resized images after creating the ZIP file
                 foreach ($this->availableWidths as $width) {
-                    $resizedImg = storage_path('app/public/photos_edited/' . $width . '_' . $this->nom);
-                    if (file_exists($resizedImg)) {
-                        unlink($resizedImg);
-                    }
+                    $resizedImg = storage_path('app/public/photos_edited/' . $this->nom . '-' .$width.'w.'.$this->extension);
+                    if (file_exists($resizedImg)) { unlink($resizedImg);}
                 }
-            
+    
+                $this->download = false;
                 return response()->download($zipFilePath)->deleteFileAfterSend(true);
-           }else{
-                foreach( $this->availableWidths as $width){
-                    return response()->download(
-                        storage_path('app/public/photos_edited/'.$width.'_'.$this->nom));
+            } else {
+                foreach ($this->availableWidths as $width) {
+                    $resizedImageName = $this->nom . '-' . $width . 'w.' . $this->extension;
+
+                    $this->download = false;
+                    return response()->download(storage_path('app/public/photos_edited/' . $resizedImageName))->deleteFileAfterSend(true);
                 }
             }
         }
     }
+    
 
     public function render()
     {
